Fixed a bug in the Rates Count service and added the Readme File
The original Rates Count service contained a bug in which the number
being returned by the handler was being treated as a status code rather
than a body value. I fixed this bug and also drafted the Readme file. As
an aside, note that I also deleted the git directory located in the
ElucidSolutions directory. This directory is no longer needed now that
the front and back-ends have been merged into a single package.